Fixed the drops
You tried to cast a Collection<ItemStack> to ItemStack :3. I thought I
could do .get(0),but apparently not.
So I used .toArray() to create an array (with the default one being an
array that drops AIR) and get the first one. Though it might be that the
dropItem(location, AIR stack) might throw you an error